traveler/requirements.txt
amqp==2.6.0
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5.2
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.4.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
click-man==0.3.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
defusedxml==0.7.1
Django==3.0
django-cas-ng==4.1.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-ckeditor==6.1.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-js-asset==1.2.2
django-mama-cas==2.4.0
django-markdownx==3.0.1
django-mdeditor==0.1.18
django-mysql==3.9.0
django-object-tool==0.0.10
django-simpleui==3.9
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djangorestframework-xml==2.0.0
dnspython==1.16.0
drf-access-policy==0.5.1
et-xmlfile==1.1.0
eventlet==0.25.2
future==0.18.2
greenlet==0.4.16
httpie==2.0.0
idna==2.8
importlib-metadata==3.1.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
jsonfield==3.1.0
kiwisolver==1.3.1
kombu==4.6.11
lazy-object-proxy==1.4.3
livereload==2.6.1
lxml==4.5.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib
mccabe==0.6.1
mkdocs==1.0.4
monotonic==1.5
mysqlclient
Naked==0.1.31
numpy
openpyxl==3.0.7
optionaldict==0.1.2
Pillow==7.0.0
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.7
python-cas==1.5.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
redis==3.5.3
requests==2.22.0
shellescape==3.8.1
six==1.14.0
soupsieve==2.2.1
sqlparse==0.3.0
tornado==6.0.3
typed-ast==1.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wechat-django==0.3.3
wechatpy==1.8.15
wrapt==1.11.2
xmltodict==0.12.0
zipp==3.4.0
